refactor to load components
