+++ tif2blender/load_components/__init__.py
@@ -1,0 +1,5 @@
+from .load_generic import *
+from .load_axes import *
+from .load_labelmask import *
+from .load_volume import *
+from .load_surfaces import *
