+++ tif2blender/load_components/load_surfaces.py
@@ -1,0 +1,3 @@
+import bpy
+
+from .load_generic import init_holder
